Fix assumptions about provided context
The context provided by Ariadne defaults to a dictionary containing a
"request" entry. This code made the assumption that the info.context
object was always a request. As such, the tests were all correctly
incorrect. This modifies that assumption so the code can be used within
Django + Channels with authentication.