Rework the login_required decorator to work with Starlette
Seeing startlette.requests.Request make it to this point when using
Ariadne with Channels. Checking against Django's `HttpRequest` means
this couldn't handle those requests. This moves to more of a duck-typing
model so anything that quacks can be passed in.

Side note: I _hate_ removing test cases, but these weren't running as
expected. I could invert the assertions inside the supposedly wrapped
functions and not trigger failures. Given that, the test was providing a
false sense of security so I just removed it.